finishing auth
