+++ AreaApi/src/routers/services/services.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const Services = require('../../models/Service')
+const Services = require('../../models/Service.model')
 // const Actions = require('../../models/Action')
 // const Reactions = require('../../models/Reaction')
